test + manual identity module
